Refactored imports for pyearthtools/training/wrapper Added a SimpleModel class for notebooks / worked examples Added .nc4 file extension to recognised types Improved cache file finding code
