+++ packages/data/src/pyearthtools/data/indexes/_indexes.py
@@ -201,14 +201,15 @@ class FileSystemIndex(Index, metaclass=ABCMeta):
                 Loaded Data
         """
         try:
+
             load_index = self.search(*args)
 
             cache = getattr(self, 'object_cache', {})
-            cached_result = cache.get(load_index, None)
+            cached_result = cache.get(str(load_index), None)
             result = cached_result or self.load(load_index, **kwargs)
 
             if getattr(self, 'cache_last_used', True):
-                self.object_cache = {load_index: result}
+                self.object_cache = {str(load_index): result}
 
             return result
         except Exception as e:
